Deleted Wrapped PCA transform
