+++ src/Microsoft.ML.PCA/PcaTransform.cs
@@ -1,4 +1,4 @@
-// Licensed to the .NET Foundation under one or more agreements.
+﻿// Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
@@ -15,11 +15,21 @@ using Microsoft.ML.Runtime.Internal.CpuMath;
 using Microsoft.ML.Runtime.Internal.Utilities;
 using Microsoft.ML.Runtime.Model;
 using Microsoft.ML.Runtime.Numeric;
+using Microsoft.ML.Core.Data;
+using Microsoft.ML.StaticPipe;
+using Microsoft.ML.StaticPipe.Runtime;
+using System.Collections.Generic;
 
-[assembly: LoadableClass(PcaTransform.Summary, typeof(PcaTransform), typeof(PcaTransform.Arguments), typeof(SignatureDataTransform),
+[assembly: LoadableClass(PcaTransform.Summary, typeof(IDataTransform), typeof(PcaTransform), typeof(PcaTransform.Arguments), typeof(SignatureDataTransform),
     PcaTransform.UserName, PcaTransform.LoaderSignature, PcaTransform.ShortName)]
 
-[assembly: LoadableClass(PcaTransform.Summary, typeof(PcaTransform), null, typeof(SignatureLoadDataTransform),
+[assembly: LoadableClass(PcaTransform.Summary, typeof(IDataTransform), typeof(PcaTransform), null, typeof(SignatureLoadDataTransform),
+    PcaTransform.UserName, PcaTransform.LoaderSignature)]
+
+[assembly: LoadableClass(PcaTransform.Summary, typeof(PcaTransform), null, typeof(SignatureLoadModel),
+    PcaTransform.UserName, PcaTransform.LoaderSignature)]
+
+[assembly: LoadableClass(typeof(IRowMapper), typeof(PcaTransform), null, typeof(SignatureLoadRowMapper),
     PcaTransform.UserName, PcaTransform.LoaderSignature)]
 
 [assembly: LoadableClass(typeof(void), typeof(PcaTransform), null, typeof(SignatureEntryPointModule), PcaTransform.LoaderSignature)]
@@ -27,7 +37,7 @@ using Microsoft.ML.Runtime.Numeric;
 namespace Microsoft.ML.Runtime.Data
 {
     /// <include file='doc.xml' path='doc/members/member[@name="PCA"]/*' />
-    public sealed class PcaTransform : OneToOneTransformBase
+    public sealed class PcaTransform : OneToOneTransformerBase
     {
         internal static class Defaults
         {
@@ -98,6 +108,37 @@ namespace Microsoft.ML.Runtime.Data
             }
         }
 
+        public sealed class ColumnInfo
+        {
+            public readonly string Input;
+            public readonly string Output;
+            public readonly string WeightColumn;
+            public readonly int Rank;
+            public readonly int Oversampling;
+            public readonly bool Center;
+            public readonly int? Seed;
+
+            /// <summary>
+            /// Describes how the transformer handles one column pair.
+            /// </summary>
+            public ColumnInfo(string input,
+                              string output,
+                              string weightColumn = PcaTransform.Defaults.WeightColumn,
+                              int rank = PcaTransform.Defaults.Rank,
+                              int overSampling = PcaTransform.Defaults.Oversampling,
+                              bool center = PcaTransform.Defaults.Center,
+                              int? seed = null)
+            {
+                Input = input;
+                Output = output;
+                WeightColumn = weightColumn;
+                Rank = rank;
+                Oversampling = overSampling;
+                Center = center;
+                Seed = seed;
+            }
+        }
+
         private sealed class TransformInfo
         {
             public readonly int Dimension;
@@ -106,14 +147,14 @@ namespace Microsoft.ML.Runtime.Data
             public Float[][] Eigenvectors;
             public Float[] MeanProjected;
 
-            public TransformInfo(Column item, Arguments args, int d)
+            public TransformInfo(int rank, int dim)
             {
-                Dimension = d;
-                Rank = item.Rank ?? args.Rank;
-                Contracts.CheckUserArg(0 < Rank && Rank <= Dimension, nameof(item.Rank), "Rank must be positive, and at most the dimension of untransformed data");
+                Dimension = dim;
+                Rank = rank;
+                Contracts.CheckUserArg(0 < Rank && Rank <= Dimension, nameof(Rank), "Rank must be positive, and at most the dimension of untransformed data");
             }
 
-            public TransformInfo(ModelLoadContext ctx, int colValueCount)
+            public TransformInfo(ModelLoadContext ctx)
             {
                 Contracts.AssertValue(ctx);
 
@@ -126,8 +167,6 @@ namespace Microsoft.ML.Runtime.Data
                 // Float[]: MeanProjected
 
                 Dimension = ctx.Reader.ReadInt32();
-                Contracts.CheckDecode(Dimension == colValueCount);
-
                 Rank = ctx.Reader.ReadInt32();
                 Contracts.CheckDecode(0 < Rank && Rank <= Dimension);
 
@@ -198,54 +237,55 @@ namespace Microsoft.ML.Runtime.Data
         }
 
         // These are parallel to Infos.
-        private readonly ColumnType[] _types;
+        private readonly ColumnType[] _outputColumnTypes;
         private readonly TransformInfo[] _transformInfos;
-
-        private readonly int[] _oversampling;
-        private readonly bool[] _center;
-        private readonly int[] _weightColumnIndex;
+        private readonly int[] _weightColumnIndices;
+        private readonly int[] _inputColumnIndices;
+        private readonly ColumnType[] _inputColumnTypes;
+        private readonly int _numColumns;
 
         private const string RegistrationName = "Pca";
 
         /// <summary>
         /// Public constructor corresponding to SignatureDataTransform.
         /// </summary>
-        public PcaTransform(IHostEnvironment env, Arguments args, IDataView input)
-            : base(env, RegistrationName, Contracts.CheckRef(args, nameof(args)).Column,
-                input, TestIsFloatItem)
+        public PcaTransform(IHostEnvironment env, IDataView input, ColumnInfo[] columns)
+            : base(Contracts.CheckRef(env, nameof(env)).Register(nameof(PcaTransform)), GetColumnPairs(columns))
         {
-            Host.AssertNonEmpty(Infos);
-            Host.Assert(Infos.Length == Utils.Size(args.Column));
+            Host.AssertNonEmpty(ColumnPairs);
 
-            _transformInfos = new TransformInfo[args.Column.Length];
-            _oversampling = new int[args.Column.Length];
-            _center = new bool[args.Column.Length];
-            _weightColumnIndex = new int[args.Column.Length];
-            for (int i = 0; i < _transformInfos.Length; i++)
+            _numColumns = ColumnPairs.Length;
+            _transformInfos = new TransformInfo[_numColumns];
+            _weightColumnIndices = new int[_numColumns];
+            _inputColumnIndices = new int[_numColumns];
+            _inputColumnTypes = new ColumnType[_numColumns];
+
+            for (int i = 0; i < _numColumns; i++)
             {
-                Host.Check(Infos[i].TypeSrc.VectorSize > 1, "Pca transform can only be applied to columns with known dimensionality greater than 1");
-                _transformInfos[i] = new TransformInfo(args.Column[i], args, Infos[i].TypeSrc.ValueCount);
-                _center[i] = args.Column[i].Center ?? args.Center;
-                _oversampling[i] = args.Column[i].Oversampling ?? args.Oversampling;
-                Host.CheckUserArg(_oversampling[i] >= 0, nameof(args.Oversampling), "Oversampling must be non-negative");
-                _weightColumnIndex[i] = -1;
-                var weightColumn = args.Column[i].WeightColumn ?? args.WeightColumn;
+                var col = columns[i];
+                // Base class has checked existence of input columns
+                input.Schema.TryGetColumnIndex(col.Input, out _inputColumnIndices[i]);
+                _inputColumnTypes[i] = input.Schema[_inputColumnIndices[i]].Type;
+                ValidatePcaInput(Host, col.Input, _inputColumnTypes[i]);
+                _transformInfos[i] = new TransformInfo(col.Rank, _inputColumnTypes[i].ValueCount);
+                Host.CheckUserArg(col.Oversampling >= 0, nameof(col.Oversampling), "Oversampling must be non-negative");
+                _weightColumnIndices[i] = -1;
+                var weightColumn = col.WeightColumn;
                 if (weightColumn != null)
                 {
-                    if (!Source.Schema.TryGetColumnIndex(weightColumn, out _weightColumnIndex[i]))
+                    if (!input.Schema.TryGetColumnIndex(weightColumn, out _weightColumnIndices[i]))
                         throw Host.Except("weight column '{0}' does not exist", weightColumn);
-                    var type = Source.Schema.GetColumnType(_weightColumnIndex[i]);
-                    Host.CheckUserArg(type == NumberType.Float, nameof(args.WeightColumn));
+                    var type = input.Schema.GetColumnType(_weightColumnIndices[i]);
+                    Host.CheckUserArg(type == NumberType.Float, nameof(weightColumn));
                 }
             }
 
-            Train(args, _transformInfos, input);
-
-            _types = InitColumnTypes();
+            Train(columns, _transformInfos, input);
+            _outputColumnTypes = InitColumnTypes();
         }
 
-        private PcaTransform(IHost host, ModelLoadContext ctx, IDataView input)
-            : base(host, ctx, input, TestIsFloatItem)
+        private PcaTransform(IHost host, ModelLoadContext ctx)
+         : base(host, ctx)
         {
             Host.AssertValue(ctx);
 
@@ -253,27 +293,54 @@ namespace Microsoft.ML.Runtime.Data
             // <prefix handled in static Create method>
             // <base>
             // transformInfos
-            Host.AssertNonEmpty(Infos);
-            _transformInfos = new TransformInfo[Infos.Length];
-            for (int i = 0; i < Infos.Length; i++)
-                _transformInfos[i] = new TransformInfo(ctx, Infos[i].TypeSrc.ValueCount);
-            _types = InitColumnTypes();
+            Host.AssertNonEmpty(ColumnPairs);
+            _numColumns = ColumnPairs.Length;
+            _transformInfos = new TransformInfo[_numColumns];
+            for (int i = 0; i < _numColumns; i++)
+                _transformInfos[i] = new TransformInfo(ctx);
+            _outputColumnTypes = InitColumnTypes();
         }
 
-        public static PcaTransform Create(IHostEnvironment env, ModelLoadContext ctx, IDataView input)
+        // Factory method for SignatureLoadDataTransform.
+        private static IDataTransform Create(IHostEnvironment env, ModelLoadContext ctx, IDataView input)
+            => Create(env, ctx).MakeDataTransform(input);
+
+        // Factory method for SignatureLoadRowMapper.
+        private static IRowMapper Create(IHostEnvironment env, ModelLoadContext ctx, ISchema inputSchema)
+            => Create(env, ctx).MakeRowMapper(inputSchema);
+
+        // Factory method for SignatureDataTransform.
+        private static IDataTransform Create(IHostEnvironment env, Arguments args, IDataView input)
         {
             Contracts.CheckValue(env, nameof(env));
-            var h = env.Register(RegistrationName);
-            h.CheckValue(ctx, nameof(ctx));
-            h.CheckValue(input, nameof(input));
-            ctx.CheckAtModel(GetVersionInfo());
+            env.CheckValue(args, nameof(args));
+            env.CheckValue(input, nameof(input));
+            env.CheckValue(args.Column, nameof(args.Column));
+            var cols = args.Column.Select(item => new ColumnInfo(
+                        item.Source,
+                        item.Name,
+                        item.WeightColumn,
+                        item.Rank ?? args.Rank,
+                        item.Oversampling ?? args.Oversampling,
+                        item.Center ?? args.Center,
+                        item.Seed ?? args.Seed)).ToArray();
+            return new PcaTransform(env, input, cols).MakeDataTransform(input);
+        }
 
-            // *** Binary format ***
-            // int: sizeof(Float)
-            // <remainder handled in ctors>
-            int cbFloat = ctx.Reader.ReadInt32();
-            h.CheckDecode(cbFloat == sizeof(Float));
-            return h.Apply("Loading Model", ch => new PcaTransform(h, ctx, input));
+        // Factory method for SignatureLoadModel.
+        private static PcaTransform Create(IHostEnvironment env, ModelLoadContext ctx)
+        {
+            Contracts.CheckValue(env, nameof(env));
+            var host = env.Register(nameof(PcaTransform));
+
+            host.CheckValue(ctx, nameof(ctx));
+            ctx.CheckAtModel(GetVersionInfo());
+            if (ctx.Header.ModelVerWritten == 0x00010001)
+            {
+                int cbFloat = ctx.Reader.ReadInt32();
+                env.CheckDecode(cbFloat == sizeof(float));
+            }
+            return new PcaTransform(host, ctx);
         }
 
         public override void Save(ModelSaveContext ctx)
@@ -287,23 +354,28 @@ namespace Microsoft.ML.Runtime.Data
             // <base>
             // transformInfos
             ctx.Writer.Write(sizeof(Float));
-            SaveBase(ctx);
+            SaveColumns(ctx);
             for (int i = 0; i < _transformInfos.Length; i++)
                 _transformInfos[i].Save(ctx);
         }
 
-        private void Train(Arguments args, TransformInfo[] transformInfos, IDataView trainingData)
+        private static (string input, string output)[] GetColumnPairs(ColumnInfo[] columns)
         {
-            var y = new Float[transformInfos.Length][][];
-            var omega = new Float[transformInfos.Length][][];
-            var mean = new Float[transformInfos.Length][];
+            //Contracts.CheckValue(columns, nameof(columns));
+            return columns.Select(x => (x.Input, x.Output)).ToArray();
+        }
 
-            var oversampledRank = new int[transformInfos.Length];
+        private void Train(ColumnInfo[] columns, TransformInfo[] transformInfos, IDataView trainingData)
+        {
+            var y = new Float[_numColumns][][];
+            var omega = new Float[_numColumns][][];
+            var mean = new Float[_numColumns][];
+            var oversampledRank = new int[_numColumns];
             var rnd = Host.Rand;
             Double totalMemoryUsageEstimate = 0;
-            for (int iinfo = 0; iinfo < transformInfos.Length; iinfo++)
+            for (int iinfo = 0; iinfo < _numColumns; iinfo++)
             {
-                oversampledRank[iinfo] = Math.Min(transformInfos[iinfo].Rank + _oversampling[iinfo], transformInfos[iinfo].Dimension);
+                oversampledRank[iinfo] = Math.Min(transformInfos[iinfo].Rank + columns[iinfo].Oversampling, transformInfos[iinfo].Dimension);
 
                 //exact: (size of the 2 big matrices + other minor allocations) / (2^30)
                 Double colMemoryUsageEstimate = 2.0 * transformInfos[iinfo].Dimension * oversampledRank[iinfo] * sizeof(Float) / 1e9;
@@ -313,7 +385,7 @@ namespace Microsoft.ML.Runtime.Data
                     using (var ch = Host.Start("Memory usage"))
                     {
                         ch.Info("Estimate memory usage for transforming column {1}: {0:G2} GB. If running out of memory, reduce rank and oversampling factor.",
-                            colMemoryUsageEstimate, Infos[iinfo].Name);
+                            colMemoryUsageEstimate, ColumnPairs[iinfo].input);
                     }
                 }
 
@@ -329,7 +401,7 @@ namespace Microsoft.ML.Runtime.Data
                     }
                 }
 
-                if (_center[iinfo])
+                if (columns[iinfo].Center)
                     mean[iinfo] = new Float[transformInfos[iinfo].Dimension];
             }
             if (totalMemoryUsageEstimate > 2)
@@ -386,7 +458,7 @@ namespace Microsoft.ML.Runtime.Data
         //Note that the covariance matrix is not computed explicitly
         private void Project(IDataView trainingData, Float[][] mean, Float[][][] omega, Float[][][] y, TransformInfo[] transformInfos)
         {
-            Host.Assert(mean.Length == omega.Length && omega.Length == y.Length && y.Length == Infos.Length);
+            Host.Assert(mean.Length == omega.Length && omega.Length == y.Length && y.Length == _numColumns);
             for (int i = 0; i < omega.Length; i++)
                 Contracts.Assert(omega[i].Length == y[i].Length);
 
@@ -399,37 +471,37 @@ namespace Microsoft.ML.Runtime.Data
 
             bool[] center = Enumerable.Range(0, mean.Length).Select(i => mean[i] != null).ToArray();
 
-            Double[] totalColWeight = new Double[Infos.Length];
+            Double[] totalColWeight = new Double[_numColumns];
 
-            bool[] activeColumns = new bool[Source.Schema.ColumnCount];
-            for (int iinfo = 0; iinfo < Infos.Length; iinfo++)
+            bool[] activeColumns = new bool[trainingData.Schema.ColumnCount];
+            for (int iinfo = 0; iinfo < _numColumns; iinfo++)
             {
-                activeColumns[Infos[iinfo].Source] = true;
-                if (_weightColumnIndex[iinfo] >= 0)
-                    activeColumns[_weightColumnIndex[iinfo]] = true;
+                activeColumns[_inputColumnIndices[iinfo]] = true;
+                if (_weightColumnIndices[iinfo] >= 0)
+                    activeColumns[_weightColumnIndices[iinfo]] = true;
             }
+
             using (var cursor = trainingData.GetRowCursor(col => activeColumns[col]))
             {
-                var weightGetters = new ValueGetter<Float>[Infos.Length];
-                var columnGetters = new ValueGetter<VBuffer<Float>>[Infos.Length];
-                for (int iinfo = 0; iinfo < Infos.Length; iinfo++)
+                var weightGetters = new ValueGetter<Float>[_numColumns];
+                var columnGetters = new ValueGetter<VBuffer<Float>>[_numColumns];
+                for (int iinfo = 0; iinfo < _numColumns; iinfo++)
                 {
-                    if (_weightColumnIndex[iinfo] >= 0)
-                        weightGetters[iinfo] = cursor.GetGetter<Float>(_weightColumnIndex[iinfo]);
-                    columnGetters[iinfo] = cursor.GetGetter<VBuffer<Float>>(Infos[iinfo].Source);
+                    if (_weightColumnIndices[iinfo] >= 0)
+                        weightGetters[iinfo] = cursor.GetGetter<Float>(_weightColumnIndices[iinfo]);
+                    columnGetters[iinfo] = cursor.GetGetter<VBuffer<Float>>(_inputColumnIndices[iinfo]);
                 }
 
                 var features = default(VBuffer<Float>);
                 while (cursor.MoveNext())
                 {
-                    for (int iinfo = 0; iinfo < Infos.Length; iinfo++)
+                    for (int iinfo = 0; iinfo < _numColumns; iinfo++)
                     {
-                        Contracts.Check(Infos[iinfo].TypeSrc.IsVector && Infos[iinfo].TypeSrc.ItemType.IsNumber,
+                        Contracts.Check(_inputColumnTypes[iinfo].IsVector && _inputColumnTypes[iinfo].ItemType.IsNumber,
                             "PCA transform can only be performed on numeric columns of dimension > 1");
 
                         Float weight = 1;
-                        if (weightGetters[iinfo] != null)
-                            weightGetters[iinfo](ref weight);
+                        weightGetters[iinfo]?.Invoke(ref weight);
                         columnGetters[iinfo](ref features);
 
                         if (FloatUtils.IsFinite(weight) && weight >= 0 && (features.Count == 0 || FloatUtils.IsFinite(features.Values, features.Count)))
@@ -445,13 +517,13 @@ namespace Microsoft.ML.Runtime.Data
                     }
                 }
 
-                for (int iinfo = 0; iinfo < Infos.Length; iinfo++)
+                for (int iinfo = 0; iinfo < _numColumns; iinfo++)
                 {
                     if (totalColWeight[iinfo] <= 0)
-                        throw Host.Except("Empty data in column '{0}'", Source.Schema.GetColumnName(Infos[iinfo].Source));
+                        throw Host.Except("Empty data in column '{0}'", ColumnPairs[iinfo].input);
                 }
 
-                for (int iinfo = 0; iinfo < Infos.Length; iinfo++)
+                for (int iinfo = 0; iinfo < _numColumns; iinfo++)
                 {
                     var invn = (Float)(1 / totalColWeight[iinfo]);
 
@@ -495,54 +567,108 @@ namespace Microsoft.ML.Runtime.Data
 
         private ColumnType[] InitColumnTypes()
         {
-            Host.Assert(Infos.Length == _transformInfos.Length);
-            var types = new ColumnType[Infos.Length];
-            for (int i = 0; i < _transformInfos.Length; i++)
-                types[i] = new VectorType(NumberType.Float, _transformInfos[i].Rank);
-            Metadata.Seal();
+            Host.Assert(ColumnPairs.Length == _transformInfos.Length);
+            var types = _transformInfos.Select(tInfo => new VectorType(NumberType.Float, tInfo.Rank)).ToArray();
             return types;
         }
 
-        protected override ColumnType GetColumnTypeCore(int iinfo)
+        protected override IRowMapper MakeRowMapper(ISchema schema) => new Mapper(this, Schema.Create(schema));
+
+        protected override void CheckInputColumn(ISchema inputSchema, int col, int srcCol)
         {
-            Host.Check(0 <= iinfo & iinfo < Utils.Size(_types));
-            return _types[iinfo];
+            ValidatePcaInput(Host, inputSchema.GetColumnName(srcCol), inputSchema.GetColumnType(srcCol));
         }
 
-        protected override Delegate GetGetterCore(IChannel ch, IRow input, int iinfo, out Action disposer)
+        internal static void ValidatePcaInput(IHost host, string name, ColumnType type)
         {
-            Host.AssertValueOrNull(ch);
-            Host.AssertValue(input);
-            Host.Assert(0 <= iinfo && iinfo < Infos.Length);
-            disposer = null;
+            if (!type.IsVector)
+                throw host.Except($"Pca transform can only be applied to vector columns. Column ${name} is of type ${type}");
 
-            var getSrc = GetSrcGetter<VBuffer<Float>>(input, iinfo);
-            var src = default(VBuffer<Float>);
-            var trInfo = _transformInfos[iinfo];
-            ValueGetter<VBuffer<Float>> del =
-                (ref VBuffer<Float> dst) =>
-                {
-                    getSrc(ref src);
-                    TransformFeatures(Host, ref src, ref dst, trInfo);
-                };
-            return del;
+            if (!(type.IsKnownSizeVector && type.VectorSize > 1))
+                throw host.Except($"Pca transform can only be applied to vector columns. Column ${name} is of size ${type.VectorSize}");
+
+            var itemType = type.ItemType;
+            if (!itemType.IsNumber)
+                throw host.Except($"Pca transform can only be applied to vector of numeric items. Column ${name} contains type ${itemType}");
         }
 
-        private static void TransformFeatures(IExceptionContext ectx, ref VBuffer<Float> src, ref VBuffer<Float> dst, TransformInfo transformInfo)
+        private sealed class Mapper : MapperBase
         {
-            ectx.Check(src.Length == transformInfo.Dimension);
+            private readonly ColumnType[] _outputColumnTypes;
+            // Todo: replace with ColMapNewToOld
+            private readonly ColumnType[] _inputColumnTypes;
+            private readonly int[] _inputColumnIndices;
+            private readonly PcaTransform _parent;
+            private readonly int _numColumns;
 
-            var values = dst.Values;
-            if (Utils.Size(values) < transformInfo.Rank)
-                values = new Float[transformInfo.Rank];
-
-            for (int i = 0; i < transformInfo.Rank; i++)
+            public Mapper(PcaTransform parent, Schema inputSchema)
+               : base(parent.Host.Register(nameof(Mapper)), parent, inputSchema)
             {
-                values[i] = VectorUtils.DotProductWithOffset(transformInfo.Eigenvectors[i], 0, ref src) -
-                    (transformInfo.MeanProjected == null ? 0 : transformInfo.MeanProjected[i]);
+                _parent = parent;
+                _numColumns = parent._numColumns;
+                _outputColumnTypes = parent.InitColumnTypes();
+                _inputColumnTypes = new ColumnType[_numColumns];
+                _inputColumnIndices = new int[_numColumns];
+                for (int i = 0; i < _numColumns; i++)
+                {
+                    var inputColName = _parent.ColumnPairs[i].input;
+                    if (!inputSchema.TryGetColumnIndex(inputColName, out _inputColumnIndices[i]))
+                        throw Host.ExceptSchemaMismatch(nameof(inputColName), "input", inputColName);
+                    _inputColumnTypes[i] = inputSchema[_inputColumnIndices[i]].Type;
+                    Host.Check(_inputColumnTypes[i].IsKnownSizeVector && _inputColumnTypes[i].VectorSize > 1,
+                        "Pca transform can only be applied to columns with known dimensionality greater than 1");
+                    if (_inputColumnTypes[i].VectorSize != _parent._transformInfos[i].Dimension)
+                    {
+                        var msg = $"Dimension of column ${inputColName} is ${_inputColumnTypes[i].VectorSize}, which doesn't match the expected size ${_parent._transformInfos[i].Dimension}";
+                        throw Host.Except(msg);
+                    }
+                }
+                // Ivan't comment:
+                //var getSrc = input.GetGetter<ReadOnlyMemory<char>>(ColMapNewToOld[iinfo]);
             }
 
-            dst = new VBuffer<Float>(transformInfo.Rank, values, dst.Indices);
+            public override Schema.Column[] GetOutputColumns()
+            {
+                var result = new Schema.Column[_numColumns];
+                for (int i = 0; i < _numColumns; i++)
+                    result[i] = new Schema.Column(_parent.ColumnPairs[i].output, _outputColumnTypes[i], null);
+                return result;
+            }
+
+            protected override Delegate MakeGetter(IRow input, int iinfo, out Action disposer)
+            {
+                Contracts.AssertValue(input);
+                Contracts.Assert(0 <= iinfo && iinfo < _numColumns);
+                disposer = null;
+
+                var srcGetter = input.GetGetter<VBuffer<float>>(_inputColumnIndices[iinfo]);
+                var src = default(VBuffer<float>);
+
+                ValueGetter<VBuffer<float>> dstGetter = (ref VBuffer<float> dst) =>
+                    {
+                        srcGetter(ref src);
+                        TransformFeatures(Host, ref src, ref dst, _parent._transformInfos[iinfo]);
+                    };
+
+                return dstGetter;
+            }
+
+            private static void TransformFeatures(IExceptionContext ectx, ref VBuffer<Float> src, ref VBuffer<Float> dst, TransformInfo transformInfo)
+            {
+                ectx.Check(src.Length == transformInfo.Dimension);
+
+                var values = dst.Values;
+                if (Utils.Size(values) < transformInfo.Rank)
+                    values = new Float[transformInfo.Rank];
+
+                for (int i = 0; i < transformInfo.Rank; i++)
+                {
+                    values[i] = VectorUtils.DotProductWithOffset(transformInfo.Eigenvectors[i], 0, ref src) -
+                        (transformInfo.MeanProjected == null ? 0 : transformInfo.MeanProjected[i]);
+                }
+
+                dst = new VBuffer<Float>(transformInfo.Rank, values, dst.Indices);
+            }
         }
 
         [TlcModule.EntryPoint(Name = "Transforms.PcaCalculator",
@@ -554,12 +680,154 @@ namespace Microsoft.ML.Runtime.Data
         public static CommonOutputs.TransformOutput Calculate(IHostEnvironment env, Arguments input)
         {
             var h = EntryPointUtils.CheckArgsAndCreateHost(env, "Pca", input);
-            var view = new PcaTransform(h, input, input.Data);
+            var view = PcaTransform.Create(h, input, input.Data);
             return new CommonOutputs.TransformOutput()
             {
                 Model = new TransformModel(h, view, input.Data),
                 OutputData = view
             };
         }
+    }
+
+    public sealed class PcaEstimator : IEstimator<PcaTransform>
+    {
+        private readonly IHost _host;
+        private readonly PcaTransform.ColumnInfo[] _columns;
+
+        public PcaEstimator(IHostEnvironment env, string inputColumn, string outputColumn = null,
+            string weightColumn = PcaTransform.Defaults.WeightColumn, int rank = PcaTransform.Defaults.Rank,
+            int overSampling = PcaTransform.Defaults.Oversampling, bool center = PcaTransform.Defaults.Center,
+            int? seed = null)
+            : this(env, new PcaTransform.ColumnInfo(inputColumn, outputColumn ?? inputColumn, weightColumn, rank, overSampling, center, seed))
+        {
+        }
+
+        public PcaEstimator(IHostEnvironment env, params PcaTransform.ColumnInfo[] columns)
+        {
+            Contracts.CheckValue(env, nameof(env));
+            _host = env.Register(nameof(PcaEstimator));
+            _columns = columns;
+        }
+
+        //TODO: move the dosctrings above
+        ///// <summary>
+        ///// Convinence constructor for simple one column case
+        ///// </summary>
+        ///// <include file='doc.xml' path='doc/members/member[@name="PCA"]/*' />
+        ///// <param name="env">The environment.</param>
+        ///// <param name="inputColumn">Input column to apply PCA on.</param>
+        ///// <param name="outputColumn">Output column. Null means <paramref name="inputColumn"/> is replaced.</param>
+        ///// <param name="rank">The number of components in the PCA.</param>
+        ///// <param name="advancedSettings">A delegate to apply all the advanced arguments to the algorithm.</param>
+        //public PcaEstimator(IHostEnvironment env, string inputColumn, string outputColumn = null,
+        //    int rank = PcaTransform.Defaults.Rank,
+        //    Action<PcaTransform.Arguments> advancedSettings = null)
+        //    : this(env, new[] { (inputColumn, outputColumn ?? inputColumn) }, rank, advancedSettings)
+        //{
+        //}
+
+        ///// <include file='doc.xml' path='doc/members/member[@name="PCA"]/*' />
+        ///// <param name="env">The environment.</param>
+        ///// <param name="columns">Pairs of columns to run the PCA on.</param>
+        ///// <param name="rank">The number of components in the PCA.</param>
+        ///// <param name="advancedSettings">A delegate to apply all the advanced arguments to the algorithm.</param>
+        //public PcaEstimator(IHostEnvironment env, (string input, string output)[] columns,
+        //    int rank = PcaTransform.Defaults.Rank,
+        //    Action<PcaTransform.Arguments> advancedSettings = null)
+        //{
+        //    Contracts.CheckValue(env, nameof(env));
+        //    _host = env.Register(nameof(PcaEstimator));
+
+        //    foreach (var (input, output) in columns)
+        //    {
+        //        _host.CheckUserArg(Utils.Size(input) > 0, nameof(input));
+        //        _host.CheckValue(output, nameof(output));
+        //    }
+
+        //    var args = new PcaTransform.Arguments();
+        //    args.Column = columns.Select(x => new PcaTransform.Column { Source = x.input, Name = x.output }).ToArray();
+        //    args.Rank = rank;
+        //    advancedSettings?.Invoke(args);
+        //    _columns = PcaTransform.ArgumentsToColumnInfos(args);
+        //}
+
+        public PcaTransform Fit(IDataView input) => new PcaTransform(_host, input, _columns);
+
+        public SchemaShape GetOutputSchema(SchemaShape inputSchema)
+        {
+            _host.CheckValue(inputSchema, nameof(inputSchema));
+            var result = inputSchema.Columns.ToDictionary(x => x.Name);
+            foreach (var colInfo in _columns)
+            {
+                if (!inputSchema.TryFindColumn(colInfo.Input, out var col))
+                    throw _host.ExceptSchemaMismatch(nameof(inputSchema), "input", colInfo.Input);
+
+                if (!(col.Kind == SchemaShape.Column.VectorKind.Vector && col.ItemType.IsNumber))
+                    throw _host.ExceptSchemaMismatch(nameof(inputSchema), "input", colInfo.Input);
+
+                result[colInfo.Output] = new SchemaShape.Column(colInfo.Output,
+                    SchemaShape.Column.VectorKind.Vector, NumberType.R4, false);
+            }
+
+            return new SchemaShape(result.Values);
+        }
+    }
+
+    /// <summary>
+    /// Extensions for statically typed <see cref="PcaEstimator"/>.
+    /// </summary>
+    public static class PcaEstimatorExtensions
+    {
+        private sealed class OutPipelineColumn : Vector<float>
+        {
+            public readonly Vector<float> Input;
+
+            public OutPipelineColumn(Vector<float> input, string weightColumn, int rank,
+                                     int overSampling, bool center, int? seed = null)
+                : base(new Reconciler(weightColumn, rank, overSampling, center, seed), input)
+            {
+                Input = input;
+            }
+        }
+
+        private sealed class Reconciler : EstimatorReconciler
+        {
+            private readonly PcaTransform.ColumnInfo _colInfo;
+
+            public Reconciler(string weightColumn, int rank, int overSampling, bool center, int? seed = null)
+            {
+                _colInfo = new PcaTransform.ColumnInfo(
+                    null, null, weightColumn, rank, overSampling, center, seed);
+            }
+
+            public override IEstimator<ITransformer> Reconcile(IHostEnvironment env,
+                PipelineColumn[] toOutput,
+                IReadOnlyDictionary<PipelineColumn, string> inputNames,
+                IReadOnlyDictionary<PipelineColumn, string> outputNames,
+                IReadOnlyCollection<string> usedNames)
+            {
+                // Only one column is allowed.
+                Contracts.Assert(toOutput.Length == 1);
+                var outCol = (OutPipelineColumn)toOutput[0];
+                var inputColName = inputNames[outCol.Input];
+                var outputColName = outputNames[outCol];
+                return new PcaEstimator(env, inputColName, outputColName,
+                                         _colInfo.WeightColumn, _colInfo.Rank, _colInfo.Oversampling,
+                                         _colInfo.Center, _colInfo.Seed);
+            }
+        }
+
+        // TODO: fix docstrings
+        // /// <summary>Replace current vector with its principal components. Can significantly reduce size of vector.</summary>
+        // /// <include file='doc.xml' path='doc/members/member[@name="PCA"]/*'/>
+        // /// <param name="input">The column to apply PCA to.</param>
+        // /// <param name="rank">The number of components in the PCA.</param>
+        // /// <param name="advancedSettings">A delegate to apply all the advanced arguments to the algorithm.</param>
+        public static Vector<float> ToPrincipalComponents(this Vector<float> input,
+            string weightColumn = PcaTransform.Defaults.WeightColumn,
+            int rank = PcaTransform.Defaults.Rank,
+            int overSampling = PcaTransform.Defaults.Oversampling,
+            bool center = PcaTransform.Defaults.Center,
+            int? seed = null) => new OutPipelineColumn(input, weightColumn, rank, overSampling, center, seed);
     }
 }
